+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [IMPROVEMENT] Send retry information with network requests (eg. retry_count, last_failure_status and idempotency key). See [#1991][]
+
 # 2.16.0 / 20-08-2024
 
 - [FIX] Refresh rate vital for variable refresh rate displays when over performing. See [#1973][]
@@ -745,6 +747,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1967]: https://github.com/DataDog/dd-sdk-ios/pull/1967
 [#1973]: https://github.com/DataDog/dd-sdk-ios/pull/1973
 [#1988]: https://github.com/DataDog/dd-sdk-ios/pull/1988
+[#1991]: https://github.com/DataDog/dd-sdk-ios/pull/1991
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu